third_party/blink/web_tests/external/wpt/navigation-api/ordering-and-transition/reload-intercept-reject_currententrychange-expected.txt

This is a testharness.js-based test.
[FAIL] event and promise ordering for navigation.reload() intercepted by intercept()
  assert_array_equals: expected property 3 to be "AbortSignal abort" but got "committed fulfilled" (expected array ["navigate", "currententrychange", "handler run", "AbortSignal abort", "navigateerror", "committed fulfilled", "transition.committed fulfilled", "promise microtask", "finished rejected", "transition.finished rejected"] got ["navigate", "currententrychange", "handler run", "committed fulfilled", "transition.committed fulfilled", "promise microtask", "AbortSignal abort", "navigateerror", "finished rejected", "transition.finished rejected"])
Harness: the test ran to completion.
